+++ T3_Floristeria/src/main/java/com/calvogasullmartin/t3_floristeria/LogicaInterface.java
@@ -1,8 +1,8 @@
 package com.calvogasullmartin.t3_floristeria;
 
-import com.calvogasullmartin.t3_floristeria.controladores.ControladorFuncionalInterface;
+import com.calvogasullmartin.t3_floristeria.controladores.ControladorFuncional;
 
 public interface LogicaInterface {
     
-    ControladorFuncionalInterface getControladorFuncional();
+    ControladorFuncional getControladorFuncional();
 }
