+++ T3_Floristeria/src/main/java/com/calvogasullmartin/t3_floristeria/VistaInterface.java
@@ -1,9 +1,9 @@
 package com.calvogasullmartin.t3_floristeria;
 
 import com.calvogasullmartin.t3_floristeria.controladores.ControladorFuncionalVisitor;
-import com.calvogasullmartin.t3_floristeria.controladores.ControladorFuncionalInterface;
+import com.calvogasullmartin.t3_floristeria.controladores.ControladorFuncional;
 
 public interface VistaInterface extends ControladorFuncionalVisitor{
     
-    public void interactuar(ControladorFuncionalInterface controlador);
+    public void interactuar(ControladorFuncional controlador);
 }
