+++ T3_Floristeria/src/main/java/com/calvogasullmartin/t3_floristeria/controladores/ControladorFuncional.java
@@ -1,13 +1,30 @@
 package com.calvogasullmartin.t3_floristeria.controladores;
 
-import com.calvogasullmartin.t3_floristeria.modelos.Estado;
+import com.calvogasullmartin.t3_floristeria.persistencia.DaoFactory;
 
-public abstract class ControladorFuncional extends Controlador implements ControladorFuncionalInterface{
 
-    public ControladorFuncional(Estado estado) {
-        super(estado);
-    }
+public interface ControladorFuncional {
+
+    final DaoFactory factory = DaoFactory.getFactory();
     
-    @Override
-    public abstract void aceptar(ControladorFuncionalVisitor controlador);
+    void aceptar(ControladorFuncionalVisitor controlador);
+
+    /*
+    protected Estado estado;
+
+    protected ControladorFuncional(Estado estado) {
+        this.estado = estado;
+    }
+
+    public abstract void controlar();
+
+    protected Estado getEstado() {
+        return estado;
+    }
+
+    protected void setEstado(Estado estado) {
+        assert estado != null;
+        this.estado = estado;
+    }
+    */
 }
