+++ T3_Floristeria/src/main/java/com/calvogasullmartin/t3_floristeria/controladores/ControladorFuncionalVisitor.java
@@ -2,5 +2,5 @@ package com.calvogasullmartin.t3_floristeria.controladores;
 
 public interface ControladorFuncionalVisitor {
     
-    void visitar (ArrancarAppControladorInterface arrancarAppControlador);
+    void visitar (ArrancarAppControlador arrancarAppControlador);
 }
