+++ T3_Floristeria/src/main/java/com/calvogasullmartin/t3_floristeria/controladores/locales/Controlador.java
@@ -1,4 +1,4 @@
-package com.calvogasullmartin.t3_floristeria.controladores;
+package com.calvogasullmartin.t3_floristeria.controladores.locales;
 
 import com.calvogasullmartin.t3_floristeria.modelos.Estado;
