+++ T3_Floristeria/src/main/java/com/calvogasullmartin/t3_floristeria/vistas/BienvenidaVista.java
@@ -1,22 +1,25 @@
 package com.calvogasullmartin.t3_floristeria.vistas;
 
-import com.calvogasullmartin.t3_floristeria.controladores.ArrancarAppControladorInterface;
 import com.calvogasullmartin.t3_floristeria.utils.InOut;
+import com.calvogasullmartin.t3_floristeria.controladores.ArrancarAppControlador;
+import java.io.IOException;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 
 public class BienvenidaVista {
 
     private InOut io;
-    private ArrancarAppControladorInterface controlador;
+    private ArrancarAppControlador controlador;
 
     public BienvenidaVista() {
         io = new InOut();
     }
 
     //solo interactua con un ArrncarAppControlador cualquiera
-    public void interactuar(ArrancarAppControladorInterface controlador) {        
+    public void interactuar(ArrancarAppControlador controlador) {
         assert controlador != null;
         this.controlador = controlador; //lo guardamos en atributos para no tener que pasarselo a los métodos privados
-        if (controlador.isPrimeraVez()) {            
+        if (controlador.isPrimeraVez()) {
             pedirInicializacion();
         }
         darBienvienida();
@@ -25,13 +28,22 @@ public class BienvenidaVista {
 
     private void pedirInicializacion() {
         io.writeln("Aplicación ejecutandose por primera vez ...");
-        controlador.iniciarPersistencia();
-        io.writeln("Capa de persistencia inizializada con exito");
-        new FloristeriaVista().interactuar(controlador);
+        try {
+            controlador.iniciarPersistencia();
+            io.writeln("Capa de persistencia inizializada con exito");
+            new FloristeriaVista().interactuar(controlador);
+        } catch (IOException ex) {
+            ////que hacer con la excepcion
+        }
     }
 
     private void darBienvienida() {
-        String nombre = controlador.getNombreFloristeria();
-        io.writeln("Bienvenid@ al gestor de la floristeria " + nombre + ", desarrollado por CalvoGasullMartin ");
+        String nombre;
+        try {
+            nombre = controlador.getNombreFloristeria();
+            io.writeln("Bienvenid@ al gestor de la floristeria " + nombre + ", desarrollado por CalvoGasullMartin ");
+        } catch (IOException ex) {
+            ////que hacer con la excepcion
+        }
     }
 }
