+++ T3_Floristeria/src/main/java/com/calvogasullmartin/t3_floristeria/vistas/FloristeriaVista.java
@@ -1,36 +1,43 @@
 package com.calvogasullmartin.t3_floristeria.vistas;
 
-import com.calvogasullmartin.t3_floristeria.controladores.ArrancarAppControladorInterface;
 import com.calvogasullmartin.t3_floristeria.utils.InOut;
+import com.calvogasullmartin.t3_floristeria.controladores.ArrancarAppControlador;
+import java.io.IOException;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 
 public class FloristeriaVista {
-    
+
     private InOut io;
 
-    public FloristeriaVista() {        
+    public FloristeriaVista() {
         this.io = new InOut();
     }
-        
-    public void interactuar(ArrancarAppControladorInterface controlador){
+
+    public void interactuar(ArrancarAppControlador controlador) {
         String nombre = pedirNombre();
-        controlador.iniciarFloristeria(nombre);
-        controlador.guardarUnicaFloristeria();
-        io.write("Floristeria guardada con éxito.");
+        try {
+            controlador.iniciarFloristeria(nombre);
+            controlador.guardarUnicaFloristeria();
+            io.write("Floristeria guardada con éxito.");
+        } catch (IOException ex) {
+            ////que hacer con la excepcion
+        }
     }
-    
-    private String pedirNombre(){
+
+    private String pedirNombre() {
         String nombre = null;
         boolean ok = false;
         int minNumChar = 3;
-        do{
+        do {
             nombre = io.readString("Introduce el nombre de la floristeria: ");
-            if (nombre.length() < minNumChar){
+            if (nombre.length() < minNumChar) {
                 io.writeError("de al menos 3 caracteres");
-            }else {
+            } else {
                 ok = true;
             }
-        }while(!ok);
+        } while (!ok);
         return null;
     }
-    
+
 }
